+++ src/kickstart.ts
@@ -96,14 +96,14 @@ const libraryVectors: { [x: string]: { [x: string]: number[] } } = {
 		'dos 37.44 (22.5.91)': [ LvoFlags.short, 0xF91544],
 		'graphics 37.35 (23.5.91)': [ LvoFlags.long, 0xFA8E64 ],
 		'layers 37.7 (13.3.91)': [ LvoFlags.short, 0xFC272C ],
-		//'mathieeesingbas 37.3 (9.5.91)': [ LvoFlags.short, 0xFC66BC ], // our disassembler doesn't support float instructions
+		'mathieeesingbas 37.3 (9.5.91)': [ LvoFlags.short, 0xFC66BC ], // our disassembler doesn't support float instructions
 	},
 
 	// Kickstart v2.05 r37.299 (1991)(Commodore)(A600)[!].rom
 	'87508de834dc7eb47359cede72d2e3c8a2e5d8db': {
 		'exec 37.151 (1.11.91)': [ LvoFlags.short, 0xF81FB0],
 		'expansion 37.50 (28.10.91)': [ LvoFlags.short, 0xF83CAA ],
-		//'mathieeesingbas 37.3 (9.5.91)': [ LvoFlags.short, 0xF848AC ],
+		'mathieeesingbas 37.3 (9.5.91)': [ LvoFlags.short, 0xF848AC ],
 		'dos 37.45 (21.10.91)': [ LvoFlags.short, 0xF95A40 ],
 		'graphics 37.41 (31.10.91)': [ LvoFlags.long, 0xFB2624 ],
 	},
@@ -112,7 +112,7 @@ const libraryVectors: { [x: string]: { [x: string]: number[] } } = {
 	'70033828182fffc7ed106e5373a8b89dda76faa5': {
 		'exec 39.47 (28.8.92)': [ LvoFlags.short, 0xF82280 ],
 		'expansion 39.7 (7.6.92)': [ LvoFlags.short, 0xF8378E ],
-		//'mathieeesingbas 37.3 (9.5.91)': [ LvoFlags.short, 0xF86740 ],
+		'mathieeesingbas 37.3 (9.5.91)': [ LvoFlags.short, 0xF86740 ],
 		'dos 39.23 (8.9.92)': [ LvoFlags.short, 0xF971EC ],
 		'graphics 39.89 (1.9.92)': [ LvoFlags.long, 0xFBCA7C ],
 	},
@@ -317,7 +317,6 @@ static Structures(void) {
 		}
 		//console.log(`Found library @ ${(lib.offset + this.base).toString(16)}: ${lib.name} V${lib.version} [${lib.id}] flags: ${this.data[lib.offset + 10].toString(16)}; ${vectors.length} LVOs`);
 		if(this.fdPath !== '') {
-			if(lib.name === 'mathieeesingbas.library') return; // until we can disassemble FFP instructions
 			const fdPath = path.join(this.fdPath, lib.name.replace('.library', '_lib.fd'));
 			if(fs.existsSync(fdPath)) {
 				const fd = new FD(fdPath);